pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aqi-ml-service"
version = "0.1.0"
description = "Air Quality (AQI) ML service: local + AWS training & deployment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "xgboost>=2.0.0",
  "scikit-learn>=1.5.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "typer>=0.12.3",
  "boto3>=1.34.0",
  "s3fs>=2024.3.1"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
